docs(svg-viewer): migrate story to mdx (#541)
